Function that returns the natural square root of a number

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -11,8 +11,8 @@ int helper(int n, int start, int end)
 {
 	if (start <= end)
 	{
-		int mid = start + (end - start) / 2;
-		int mid_square = mid * mid;
+		long int mid = start + (end - start) / 2;
+		long int mid_square = mid * mid;
 	
 		if (mid_square == n)
 		{
